backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
openai==1.3.7
anthropic==0.7.8
spacy==3.7.2
textblob==0.17.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
APScheduler==3.10.4
email-validator==2.1.0
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3
aiofiles==23.2.1
jinja2==3.1.2